misc/signature/requirements.txt
flask
tensorflow
gunicorn
pillow
Flask-Limiter